Add Timestamp.timestampString
Adds Timestamp.timestampString. This field does not get set until the
Timestamp is either set or retrieved as a string. It allows the Timestamp
to calculate the value when needed, but also save it for later so that it
does not have to be unecessarily generated a second time.

Adds documentation to new field, and alters documentation to AsDouble
and AsString properties.