Add checks to Timestamp.AsString getter
Adds checks to Timestamp.AsString which will make sure modulus division
is only performed when needed. This prevents odd side effects from
occuring, such as returning the timestamp "00:-35791394:NaN", which it
was returning previously. Adds a validation check to the getter to make
sure that the timestamp will not be an erroneous one.