Add Caption Alignment Button Functionality
Adds click method handlers to mainform for Caption alignment buttons.
When one of these buttons is clicked, it gets highlighted to show which
alignment the caption currently has. The reasoning for this choice is
so that the buttons can behave similarly to the alignment buttons in
Microsoft Word 2010. Icons will eventually be added in.