Remove MarkupController Class
Removes the MarkupController Class and instead changes it into a part
of the MainForm class. This was done as there is not much point in
having a distinct class for interaction that can be done in the main
form.

Renames MarkupController.cs to MainFormMarkupInteraction.cs as all
markup interaction can be kept in this file to separate it from engine
interaction and individual control interaction.