Add HapppyWordAnimation functionality
HappyWordAnimation will now create an actual set of animations. This
set of animations and text effects are not final, but they set the
pattern that the animations will follow when it comes to generating
them.

WordAnimation and its child classes now also take the caption that the
CaptionWord belongs to. This is so that the duration for animations can
be set in the constructor.